# newline token kind when lexing