--- a/crates/rslint_lexer/src/lib.rs
+++ b/crates/rslint_lexer/src/lib.rs
@@ -36,7 +36,7 @@ use state::LexerState;
 use tables::derived_property::*;
 
 pub use rslint_syntax::*;
-pub type LexerReturn = (Token, Option<Diagnostic>);
+pub type LexerReturn = (Token, Option<Box<Diagnostic>>);
 
 // Simple macro for unwinding a loop
 macro_rules! unwind_loop {
@@ -116,6 +116,58 @@ impl<'src> Lexer<'src> {
 			state: LexerState::new(),
 			returned_eof: false,
 		}
+	}
+
+	pub fn get_all(&mut self) -> (Vec<Token>, Vec<Box<Diagnostic>>) {
+		let mut tokens = vec![];
+		let mut diagnostics = vec![];
+		loop {
+			if self.cur >= self.bytes.len() {
+				if !self.returned_eof {
+					self.returned_eof = true;
+					let mut token: (Token, Option<Diagnostic>) = tok!(EOF, 0);
+					token.0.offset = self.cur as u32;
+					token.0.after_newline = self.state.after_newline;
+					tokens.push(token.0);
+				}
+				break;
+			}
+	
+			let mut token = if self.state.is_in_template() {
+				self.lex_template()
+			} else {
+				self.lex_token()
+			};
+	
+			let after_newline = match token.0.kind {
+				JsSyntaxKind::NEWLINE | JsSyntaxKind::MULTILINE_COMMENT => true,
+				JsSyntaxKind::WHITESPACE | JsSyntaxKind::COMMENT => self.state.after_newline,
+				_ => false,
+			};
+	
+			token.0.offset = self.cur as u32 - token.0.len;
+			token.0.after_newline = std::mem::replace(&mut self.state.after_newline, after_newline);
+	
+			if ![
+				JsSyntaxKind::COMMENT,
+				JsSyntaxKind::MULTILINE_COMMENT,
+				JsSyntaxKind::WHITESPACE,
+				JsSyntaxKind::NEWLINE,
+				JsSyntaxKind::TEMPLATE_CHUNK,
+			]
+			.contains(&token.0.kind)
+			{
+				self.state.update(token.0.kind);
+			}
+
+			tokens.push(token.0);
+
+			if let Some(diagnostic) = token.1 {
+ 				diagnostics.push(diagnostic);
+			}
+		}
+
+		(tokens, diagnostics)
 	}
 
 	// Bump the lexer and return the token given in
@@ -218,7 +270,7 @@ impl<'src> Lexer<'src> {
 	}
 
 	// Read a `\u{000...}` escape sequence, this expects the cur char to be the `{`
-	fn read_codepoint_escape(&mut self) -> Result<char, Diagnostic> {
+	fn read_codepoint_escape(&mut self) -> Result<char, Box<Diagnostic>> {
 		let start = self.cur + 1;
 		self.read_hexnumber();
 
@@ -230,7 +282,7 @@ impl<'src> Lexer<'src> {
                 .primary(self.cur .. invalid.len_utf8(), "");
 
 			self.cur -= 1;
-			return Err(err);
+			return Err(Box::new(err));
 		}
 
 		// Safety: We know for a fact this is in bounds because we must be on the possible char after the } at this point
@@ -255,7 +307,7 @@ impl<'src> Lexer<'src> {
 						Diagnostic::error(self.file_id, "", "invalid codepoint for unicode escape")
 							.primary(start..self.cur, "");
 
-					Err(err)
+					Err(Box::new(err))
 				}
 			}
 
@@ -268,14 +320,14 @@ impl<'src> Lexer<'src> {
 				.primary(start..self.cur, "")
 				.footer_note("Codepoints range from 0 to 0x10FFFF (1114111)");
 
-				Err(err)
+				Err(Box::new(err))
 			}
 		}
 	}
 
 	// Read a `\u0000` escape sequence, this expects the current char to be the `u`, it also does not skip over the escape sequence
 	// The pos after this method is the last hex digit
-	fn read_unicode_escape(&mut self, advance: bool) -> Result<char, Diagnostic> {
+	fn read_unicode_escape(&mut self, advance: bool) -> Result<char, Box<Diagnostic>> {
 		debug_assert_eq!(self.bytes[self.cur], b'u');
 
 		let diagnostic = Diagnostic::error(
@@ -294,13 +346,13 @@ impl<'src> Lexer<'src> {
 					if !advance {
 						self.cur -= idx + 1;
 					}
-					return Err(diagnostic);
+					return Err(Box::new(diagnostic));
 				}
 				Some(b) if !b.is_ascii_hexdigit() => {
 					if !advance {
 						self.cur -= idx + 1;
 					}
-					return Err(diagnostic);
+					return Err(Box::new(diagnostic));
 				}
 				_ => {}
 			}
@@ -327,7 +379,7 @@ impl<'src> Lexer<'src> {
 
 	// Validate a `\x00 escape sequence, this expects the current char to be the `x`, it also does not skip over the escape sequence
 	// The pos after this method is the last hex digit
-	fn validate_hex_escape(&mut self) -> Option<Diagnostic> {
+	fn validate_hex_escape(&mut self) -> Option<Box<Diagnostic>> {
 		debug_assert_eq!(self.bytes[self.cur], b'x');
 
 		let diagnostic =
@@ -339,8 +391,8 @@ impl<'src> Lexer<'src> {
 
 		for _ in 0..2 {
 			match self.next_bounded() {
-				None => return Some(diagnostic),
-				Some(b) if !(*b as u8).is_ascii_hexdigit() => return Some(diagnostic),
+				None => return Some(Box::new(diagnostic)),
+				Some(b) if !(*b as u8).is_ascii_hexdigit() => return Some(Box::new(diagnostic)),
 				_ => {}
 			}
 		}
@@ -348,7 +400,7 @@ impl<'src> Lexer<'src> {
 	}
 
 	// Validate a `\..` escape sequence and advance the lexer based on it
-	fn validate_escape_sequence(&mut self) -> Option<Diagnostic> {
+	fn validate_escape_sequence(&mut self) -> Option<Box<Diagnostic>> {
 		let cur = self.cur;
 		if let Some(escape) = self.bytes.get(self.cur + 1) {
 			match escape {
@@ -372,10 +424,11 @@ impl<'src> Lexer<'src> {
 				}
 			}
 		} else {
-			Some(Diagnostic::error(self.file_id, "", "").primary(
+			let diag = Diagnostic::error(self.file_id, "", "").primary(
 				cur..cur + 1,
 				"expected an escape sequence following a backslash, but found none",
-			))
+			);
+			Some(Box::new(diag))
 		}
 	}
 
@@ -421,7 +474,7 @@ impl<'src> Lexer<'src> {
 
 	// Consume a string literal and advance the lexer, and returning a list of errors that occurred when reading the string
 	// This could include unterminated string and invalid escape sequences
-	fn read_str_literal(&mut self) -> Option<Diagnostic> {
+	fn read_str_literal(&mut self) -> Option<Box<Diagnostic>> {
 		// Safety: this is only ever called from lex_token, which is guaranteed to be called on a char position
 		let quote = unsafe { *self.bytes.get_unchecked(self.cur) };
 		let start = self.cur;
@@ -444,7 +497,7 @@ impl<'src> Lexer<'src> {
 			.primary(self.cur..self.cur, "input ends here")
 			.secondary(start..start + 1, "string literal starts here");
 
-		Some(unterminated)
+		Some(Box::new(unterminated))
 	}
 
 	/// Returns `Some(x)` if the current position is an identifier, with the character at
@@ -621,7 +674,7 @@ impl<'src> Lexer<'src> {
 	}
 
 	#[inline]
-	fn read_zero(&mut self) -> Option<Diagnostic> {
+	fn read_zero(&mut self) -> Option<Box<Diagnostic>> {
 		// TODO: Octal literals
 		match self.bytes.get(self.cur + 1) {
 			Some(b'x') | Some(b'X') => {
@@ -687,7 +740,7 @@ impl<'src> Lexer<'src> {
 	}
 
 	#[inline]
-	fn read_hexnumber(&mut self) -> Option<Diagnostic> {
+	fn read_hexnumber(&mut self) -> Option<Box<Diagnostic>> {
 		let mut diag = None;
 		unwind_loop! {
 			match self.next() {
@@ -699,7 +752,7 @@ impl<'src> Lexer<'src> {
 	}
 
 	#[inline]
-	fn handle_numeric_separator(&mut self, radix: u8) -> Option<Diagnostic> {
+	fn handle_numeric_separator(&mut self, radix: u8) -> Option<Box<Diagnostic>> {
 		debug_assert_eq!(self.bytes[self.cur], b'_');
 
 		let err_diag = Diagnostic::error(
@@ -712,7 +765,7 @@ impl<'src> Lexer<'src> {
 		let peeked = self.bytes.get(self.cur + 1).copied();
 
 		if peeked.is_none() || !char::from(peeked.unwrap()).is_digit(radix as u32) {
-			return Some(err_diag);
+			return Some(Box::new(err_diag));
 		}
 
 		let forbidden = |c: Option<u8>| {
@@ -731,7 +784,7 @@ impl<'src> Lexer<'src> {
 		let prev = self.bytes.get(self.cur - 1).copied();
 
 		if forbidden(prev) || forbidden(peeked) {
-			return Some(err_diag);
+			return Some(Box::new(err_diag));
 		}
 
 		self.next_bounded();
@@ -741,7 +794,7 @@ impl<'src> Lexer<'src> {
 	// Read a number which does not start with 0, since that can be more things and is handled
 	// by another function
 	#[inline]
-	fn read_number(&mut self) -> Option<Diagnostic> {
+	fn read_number(&mut self) -> Option<Box<Diagnostic>> {
 		let mut diag = None;
 		unwind_loop! {
 			match self.next_bounded() {
@@ -776,7 +829,7 @@ impl<'src> Lexer<'src> {
 	}
 
 	#[inline]
-	fn read_float(&mut self) -> Option<Diagnostic> {
+	fn read_float(&mut self) -> Option<Box<Diagnostic>> {
 		let mut diag = None;
 
 		unwind_loop! {
@@ -806,7 +859,7 @@ impl<'src> Lexer<'src> {
 	}
 
 	#[inline]
-	fn read_exponent(&mut self) -> Option<Diagnostic> {
+	fn read_exponent(&mut self) -> Option<Box<Diagnostic>> {
 		if let Some(b'-') | Some(b'+') = self.bytes.get(self.cur + 1) {
 			self.next();
 		}
@@ -822,7 +875,7 @@ impl<'src> Lexer<'src> {
 	}
 
 	#[inline]
-	fn read_bindigits(&mut self) -> Option<Diagnostic> {
+	fn read_bindigits(&mut self) -> Option<Box<Diagnostic>> {
 		let mut diag = None;
 		unwind_loop! {
 			match self.next() {
@@ -834,7 +887,7 @@ impl<'src> Lexer<'src> {
 	}
 
 	#[inline]
-	fn read_octaldigits(&mut self) -> Option<Diagnostic> {
+	fn read_octaldigits(&mut self) -> Option<Box<Diagnostic>> {
 		let mut diag = None;
 		unwind_loop! {
 			match self.next() {
@@ -859,7 +912,7 @@ impl<'src> Lexer<'src> {
 
 			(
 				Token::new(JsSyntaxKind::ERROR_TOKEN, self.cur - start),
-				Some(err),
+				Some(Box::new(err)),
 			)
 		} else {
 			tok!(JS_NUMBER_LITERAL, self.cur - start)
@@ -892,7 +945,7 @@ impl<'src> Lexer<'src> {
 			)
 			.primary(0usize..1usize, "");
 
-			(Token::new(JsSyntaxKind::ERROR_TOKEN, 1), Some(err))
+			(Token::new(JsSyntaxKind::ERROR_TOKEN, 1), Some(Box::new(err)))
 		}
 	}
 
@@ -931,7 +984,7 @@ impl<'src> Lexer<'src> {
 
 				(
 					Token::new(JsSyntaxKind::COMMENT, self.cur - start),
-					Some(err),
+					Some(Box::new(err)),
 				)
 			}
 			Some(b'/') => {
@@ -1025,7 +1078,7 @@ impl<'src> Lexer<'src> {
 									}
 								},
 								_ => {
-									return (Token::new(JsSyntaxKind::JS_REGEX_LITERAL, self.cur - start), diagnostic)
+									return (Token::new(JsSyntaxKind::JS_REGEX_LITERAL, self.cur - start), diagnostic.map(Box::new))
 								}
 							}
 						}
@@ -1036,7 +1089,7 @@ impl<'src> Lexer<'src> {
 						let err = Diagnostic::error(self.file_id, "", "expected a character after a regex escape, but found none")
 							.primary(self.cur..self.cur + 1, "expected a character following this");
 
-						return (Token::new(JsSyntaxKind::JS_REGEX_LITERAL, self.cur - start), Some(err));
+						return (Token::new(JsSyntaxKind::JS_REGEX_LITERAL, self.cur - start), Some(Box::new(err)));
 					}
 				},
 				None => {
@@ -1044,7 +1097,7 @@ impl<'src> Lexer<'src> {
 						.primary(self.cur..self.cur, "...but the file ends here")
 						.secondary(start..start + 1, "a regex literal starts there...");
 
-					return (Token::new(JsSyntaxKind::JS_REGEX_LITERAL, self.cur - start), Some(err));
+					return (Token::new(JsSyntaxKind::JS_REGEX_LITERAL, self.cur - start), Some(Box::new(err)));
 				},
 				_ => {},
 			}
@@ -1327,7 +1380,7 @@ impl<'src> Lexer<'src> {
 								self.next();
 								(
 									Token::new(JsSyntaxKind::ERROR_TOKEN, self.cur - start),
-									Some(err),
+									Some(Box::new(err)),
 								)
 							}
 						}
@@ -1345,7 +1398,7 @@ impl<'src> Lexer<'src> {
 					.primary(start..self.cur + 1, "");
 					self.next();
 
-					(Token::new(JsSyntaxKind::ERROR_TOKEN, 1), Some(err))
+					(Token::new(JsSyntaxKind::ERROR_TOKEN, 1), Some(Box::new(err)))
 				}
 			}
 			QOT => {
@@ -1405,7 +1458,7 @@ impl<'src> Lexer<'src> {
 
 						(
 							Token::new(JsSyntaxKind::ERROR_TOKEN, self.cur - start),
-							Some(err),
+							Some(Box::new(err)),
 						)
 					}
 				}
@@ -1420,7 +1473,7 @@ impl<'src> Lexer<'src> {
 				.primary(start..self.cur + 1, "");
 				self.next();
 
-				(Token::new(JsSyntaxKind::ERROR_TOKEN, 1), Some(err))
+				(Token::new(JsSyntaxKind::ERROR_TOKEN, 1), Some(Box::new(err)))
 			}
 		}
 	}
@@ -1468,7 +1521,7 @@ impl<'src> Lexer<'src> {
 
 		(
 			Token::new(JsSyntaxKind::TEMPLATE_CHUNK, self.cur - start),
-			Some(err),
+			Some(Box::new(err)),
 		)
 	}
 }