--- a/crates/rslint_lexer/src/state.rs
+++ b/crates/rslint_lexer/src/state.rs
@@ -7,6 +7,7 @@ pub(crate) struct LexerState {
 	pub(crate) prev: Option<JsSyntaxKind>,
 	pub(crate) had_linebreak: bool,
 	pub(crate) ctx: Vec<Context>,
+	pub(crate) after_newline: bool, 
 }
 
 impl LexerState {
@@ -15,6 +16,7 @@ impl LexerState {
 			expr_allowed: true,
 			prev: None,
 			had_linebreak: false,
+			after_newline: false,
 			ctx: vec![Context::BraceStmt],
 		}
 	}