--- a/crates/rslint_lexer/src/state.rs
+++ b/crates/rslint_lexer/src/state.rs
@@ -7,7 +7,7 @@ pub(crate) struct LexerState {
 	pub(crate) prev: Option<JsSyntaxKind>,
 	pub(crate) had_linebreak: bool,
 	pub(crate) ctx: Vec<Context>,
-	pub(crate) after_newline: bool, 
+	pub(crate) after_newline: bool,
 }
 
 impl LexerState {