… Cargo.lock
… crates/rome_rowan/Cargo.toml
… crates/rslint_parser/Cargo.toml
… crates/rslint_parser/src/event.rs
… crates/rslint_parser/src/lossless_tree_sink.rs
… crates/rslint_parser/src/lossy_tree_sink.rs
… crates/rslint_parser/src/parse.rs
… crates/rslint_parser/src/parser.rs
… crates/rslint_parser/src/parser/parsed_syntax.rs
… crates/rslint_parser/src/syntax/class.rs
… crates/rslint_parser/src/syntax/decl.rs
… crates/rslint_parser/src/syntax/expr.rs
… crates/rslint_parser/src/syntax/js_parse_error.rs
… crates/rslint_parser/src/syntax/module.rs
… crates/rslint_parser/src/syntax/pattern.rs
… crates/rslint_parser/src/syntax/program.rs
… crates/rslint_parser/src/syntax/stmt.rs
… crates/rslint_parser/src/syntax/typescript.rs
… crates/rslint_parser/src/syntax/util.rs
… crates/rslint_parser/src/tests.rs
… crates/rslint_parser/src/token_source.rs
… crates/rslint_parser/test_data/inline/err/binding_identifier_invalid.rast
… crates/rslint_parser/test_data/inline/err/debugger_stmt.rast
… crates/rslint_parser/test_data/inline/err/directives_err.rast
… crates/rslint_parser/test_data/inline/err/double_label.rast
… crates/rslint_parser/test_data/inline/err/export_decl_not_top_level.rast
… crates/rslint_parser/test_data/inline/err/getter_class_no_body.rast
… crates/rslint_parser/test_data/inline/err/import_decl_not_top_level.rast
… crates/rslint_parser/test_data/inline/err/invalid_method_recover.rast
… crates/rslint_parser/test_data/inline/err/method_getter_err.rast