PS: This PR is outdated. See https://github.com/rome/tools/issues/1901.

We were treating whitespace and newline as ```WHITESPACE```. Which is fine, but it was generating the problem that we were peeking into the ```WHITESPACE``` token text to find new lines. Now we don't need to do this anymore, which is making the parser a lit bit faster.

# Example

Before
```rust
assert_lex! {
	"\n ",
	WHITESPACE:2
}
```

After
```rust
assert_lex! {
	"\n ",
	NEWLINE:1
	WHITESPACE:1
}
```

Or with ```\r\n```

Before 
```rust
assert_lex! {
	" \r\n",
	WHITESPACE:1
	WHITESPACE:2
}
```

After
```rust
assert_lex! {
	" \r\n",
	WHITESPACE:1
	NEWLINE:2
}
```

# Multiple lines

Today, when we find multiple lines we merge them together in one ```WHITESPACE``` token. Now we are splitting them in multiple ```NEWLINE``` tokens, for example:

```rust

#[test]
fn all_whitespace() {
	...
	assert_lex! {
		"\n\n",
		NEWLINE:1
		NEWLINE:1
	}
        ...
}
```

# Performance

Performance gains came from:
1 - we don't iterate the token text for newlines anymore;
2 - we don't create token_offset_pair and tokens_after_linebreaks; and obviously, we don't search them anymore.
2.1 - We pre-calculate token offset and if it is the next non-trivia token after newlines;
3 - We don't copy needlessly tokens when parsing;
4 - Token struct is smaller (12 bytes versus 16 bytes), even though we have "more data" now.
5 - Faster allocator

|lib|CPU Before|CPU After|Memory Before| Memory After|
--- | --- | ---| ---| ---|
|jquery|best 25ms|average 16.381ms|  Max Bytes: 8.71 MB; Total Bytes: 16.65 MB | Max Bytes: 8.57 MB; Total Bytes: 16.65 MB |
|threejs|best 117ms|average 89.533ms| Max Bytes: 35.75 MB; Total Bytes: 562.57 MB | Max Bytes: 35.04 MB; Total Bytes: 562.55 MB |

Before
```
Total Time: 43.6018ms
Total Time: 98.2213ms
Total Time: 2.6924ms
Total Time: 65.8382ms
Total Time: 236.5217ms
Total Time: 10.5943ms
Total Time: 146.1323ms
Total Time: 142.3762ms
Total Time: 175.1149ms
Total Time: 268.3024ms
```

Output is different but the libs follow the same order

After(Windows)
```
[jquery.min.js] Total Time: 16.484ms (tokenization: 1.9288ms, parsing: 7.568ms, tree_sink: 6.9872ms)
[vue.global.prod.js] Total Time: 36.8386ms (tokenization: 4.1165ms, parsing: 19.4657ms, tree_sink: 13.2564ms)
[react.production.min.js] Total Time: 1.2815ms (tokenization: 206.9┬╡s, parsing: 549.2┬╡s, tree_sink: 525.4┬╡s)  
[react-dom.production.min.js] Total Time: 26.9426ms (tokenization: 2.5768ms, parsing: 13.0159ms, tree_sink: 11.3499ms)
[compiler.js] Total Time: 90.265ms (tokenization: 15.4077ms, parsing: 30.0471ms, tree_sink: 44.8102ms)
[dojo.js] Total Time: 4.714ms (tokenization: 791.5┬╡s, parsing: 1.4962ms, tree_sink: 2.4263ms)
[d3.min.js] Total Time: 57.3223ms (tokenization: 6.2511ms, parsing: 26.2195ms, tree_sink: 24.8517ms)
[pixi.min.js] Total Time: 66.207ms (tokenization: 7.7532ms, parsing: 29.5057ms, tree_sink: 28.9481ms)
[three.min.js] Total Time: 77.0848ms (tokenization: 8.9038ms, parsing: 35.8389ms, tree_sink: 32.3421ms)
[tex-chtml-full.js] Total Time: 166.347ms (tokenization: 17.9924ms, parsing: 64.6204ms, tree_sink: 83.7342ms)
[lodash.js] Total Time: 24.9849ms (tokenization: 4.2597ms, parsing: 8.6357ms, tree_sink: 12.0895ms)
```

After (WSL)
```
[jquery.min.js] Total Time: 12.199781ms (tokenization: 1.682611ms, parsing: 4.70618ms, tree_sink: 5.81099ms)     
[vue.global.prod.js] Total Time: 26.576707ms (tokenization: 3.532758ms, parsing: 12.469799ms, tree_sink: 10.57415ms)
[react.production.min.js] Total Time: 838.385µs (tokenization: 137.062µs, parsing: 288.282µs, tree_sink: 413.041µs)
[react-dom.production.min.js] Total Time: 17.357789ms (tokenization: 3.02271ms, parsing: 5.770853ms, tree_sink: 8.564226ms)
[compiler.js] Total Time: 66.397749ms (tokenization: 12.879798ms, parsing: 18.502785ms, tree_sink: 35.015166ms)  
[dojo.js] Total Time: 3.155883ms (tokenization: 571.429µs, parsing: 1.01304ms, tree_sink: 1.571414ms)
[d3.min.js] Total Time: 40.242459ms (tokenization: 4.844913ms, parsing: 15.089494ms, tree_sink: 20.308052ms)     
[pixi.min.js] Total Time: 48.257906ms (tokenization: 7.138972ms, parsing: 17.613233ms, tree_sink: 23.505701ms)   
[three.min.js] Total Time: 60.235828ms (tokenization: 9.597417ms, parsing: 24.153288ms, tree_sink: 26.485123ms)  
[tex-chtml-full.js] Total Time: 103.596197ms (tokenization: 14.693265ms, parsing: 39.13065ms, tree_sink: 49.772282ms)
[lodash.js] Total Time: 15.36596ms (tokenization: 3.387414ms, parsing: 4.313899ms, tree_sink: 7.664647ms) 
```

The difference between Windows and WSL (I think) comes from the allocator. For example, see https://www.reddit.com/r/rust/comments/qnfcdx/very_slow_deallocation_of_boxt_on_windows.

If we use a faster allocator, we can squeeze even more performance:

After (now Windows and WSL have similar numbers)
```
[jquery.min.js] Total Time: 10.182497ms (tokenization: 1.370557ms, parsing: 3.806857ms, tree_sink: 5.005083ms)   
[vue.global.prod.js] Total Time: 18.897428ms (tokenization: 2.309826ms, parsing: 8.105518ms, tree_sink: 8.482084ms)
[react.production.min.js] Total Time: 608.614µs (tokenization: 87.132µs, parsing: 223.124µs, tree_sink: 298.358µs)
[react-dom.production.min.js] Total Time: 15.292781ms (tokenization: 1.727245ms, parsing: 5.567364ms, tree_sink: 
7.998172ms)
[compiler.js] Total Time: 61.509581ms (tokenization: 9.736649ms, parsing: 19.206972ms, tree_sink: 32.56596ms)    
[dojo.js] Total Time: 2.731947ms (tokenization: 512.855µs, parsing: 861.773µs, tree_sink: 1.357319ms)
[d3.min.js] Total Time: 36.161736ms (tokenization: 4.403448ms, parsing: 13.652004ms, tree_sink: 18.106284ms)     
[pixi.min.js] Total Time: 47.889442ms (tokenization: 6.162049ms, parsing: 19.753283ms, tree_sink: 21.97411ms)    
[three.min.js] Total Time: 59.604656ms (tokenization: 6.875095ms, parsing: 24.566829ms, tree_sink: 28.162732ms)  
[tex-chtml-full.js] Total Time: 95.775742ms (tokenization: 13.025206ms, parsing: 34.555772ms, tree_sink: 48.194764ms)
[lodash.js] Total Time: 13.902207ms (tokenization: 2.923941ms, parsing: 4.667398ms, tree_sink: 6.310868ms) 
```

It is very hard to get comparable numbers, but this is how long Chrome/Brave is taking with jquery:

![image](https://user-images.githubusercontent.com/83425/148114202-f7372ce3-d25a-4f65-a1c6-54062f9b17b6.png)